+++ lib/uchi.rb
@@ -16,7 +16,7 @@ require "uchi/plugins"
 require "uchi/repository"
 require "uchi/routes"
 require "uchi/sort_order"
-require "uchi/view"
+require "uchi/view/constants"
 require "uchi/repository/translate"
 
 module Uchi